Throw together an `Aff Unit` driver
We provide a way to interpret a component as `Aff Unit`.
Currently, it's not clear how to use this.
Things will probably flesh themselves out soon enough though.

We also provide a helper for the `eval` of a component.
This is a bit more useful for doing things like testing
(which was the impetus for this package existing).