Don't require a "child" function
Being able to write a `g ~> f` that is useful is mostly impossible.

Child queries are transformed into messages.
Messages can be transformed into parent queries.
However, messages are run inside of `HalogenM _ _ _ _ _ _ _`.
And there's no real way to have `HalogenM _ f _ _ _ _ ~> f`,
unless you really have `HalogenM _ f _ _ _ f ~> f`.
Which doesn't seem all that useful of a function to have.

The other issue is that children aren't usually created in `eval`.
The `render` of a component is where children materialize.
So, attempting to send a query to a child will never do anything.

These two issues meant that attempting to do anything with the children
was a waste of cycles.
We can eschew the conversion function (that won't generally exist),
and make `evalAff` a bit easier to use.